Добавлены на карту метки объявлений
