+++ js/main.js
@@ -1,2 +1,1 @@
-import './render-ads.js';
-import './map-initialization.js';
+import './render-map.js';
